api_client/requirements.txt
kafka-python==2.0.2
protobuf==4.24.3
grequests==0.7.0
